+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 💡 Enhancements 💡
 
+- `internal/stanza` : Export metrics from Stanza receivers (#8024)
 - `sumologicexporter`: Move validation to Config (#7936)
 - `elasticsearchexporter`: Fix crash with batch processor (#7953).
 - `splunkhecexporter`: Batch metrics payloads (#7760)